feat: resolve #5 — iOS/web support via Flutter web + GitHub Pages
- Split article_screen.dart into conditional-import trio: mobile keeps
  the existing WebView, web opens Wikipedia via url_launcher in the
  system browser and records the visit in the notebook
- Guard _autoCheckUpdate, _checkUpdates, and _downloadAndInstall in
  start_screen.dart with kIsWeb so APK update logic never runs on web
- Add web/ directory with PWA manifest (castle colours, portrait,
  start_url /Mind-mazeish/) and index.html with Apple "Add to Home
  Screen" meta tags
- Add .github/workflows/cd-web.yml: flutter build web --base-href
  /Mind-mazeish/ deployed to gh-pages via peaceiris/actions-gh-pages
