+++ web/manifest.json
@@ -1,0 +1,35 @@
+{
+  "name": "Mind Mazeish",
+  "short_name": "Mind Maze",
+  "start_url": "/Mind-mazeish/",
+  "display": "standalone",
+  "background_color": "#1A1208",
+  "theme_color": "#FF8C00",
+  "description": "Medieval castle trivia game",
+  "orientation": "portrait",
+  "prefer_related_applications": false,
+  "icons": [
+    {
+      "src": "icons/Icon-192.png",
+      "sizes": "192x192",
+      "type": "image/png"
+    },
+    {
+      "src": "icons/Icon-512.png",
+      "sizes": "512x512",
+      "type": "image/png"
+    },
+    {
+      "src": "icons/Icon-maskable-192.png",
+      "sizes": "192x192",
+      "type": "image/png",
+      "purpose": "maskable"
+    },
+    {
+      "src": "icons/Icon-maskable-512.png",
+      "sizes": "512x512",
+      "type": "image/png",
+      "purpose": "maskable"
+    }
+  ]
+}
